Split views file
